creacion de los Accesor y mutator para el modelo de detalle de factura en el precio unitario

+++ database/migrations/2025_03_11_230300_create_detalle_facturas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('factura_id')->constrained('facturas')->onDelete('cascade');
             $table->foreignId('producto_id')->constrained('productos')->onDelete('cascade');
             $table->integer('cantidad');
-            $table->decimal('precio_unitario', 12,3);
+            $table->bigInteger('precio_unitario');
             $table->timestamps();
         });
     }
